Add files via upload
This is a Jupyter Notebook file and has been run and tested in the same. All packages as mentioned in the first cell must be imported (if not done so already). All the cells must be run (Shift + Enter) in top to down approach. It might take some time in rendering the plots (depending on the resources available). A web browser view of the Jupyter Notebook has also been provided which does not require all the packages to be installed on the local machine.