Add files via upload
This is the browser view of the .ipynb file that contains all the processed code. It shows the interactive view of all the plots that have been designed based on the dataset. The browser view has the advantage of not needing python packages to be installed locally on a machine. instead it renders the plots automatically in your browser. Rendering may take some time, so be PATIENT! Also, this is the source code, so you can download it as .html file and play with it. If you want a direct browser view, I will link a html preview named "View in Browser" !